solution/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile ./solution/requirements/requirements.in
#
anyio==3.6.2
    # via starlette
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
fastapi==0.95.1
    # via -r ./solution/requirements/requirements.in
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2023.5.0
    # via huggingface-hub
h11==0.14.0
    # via uvicorn
huggingface-hub==0.14.1
    # via transformers
idna==3.4
    # via
    #   anyio
    #   requests
jinja2==3.1.2
    # via torch
markupsafe==2.1.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
numpy==1.24.3
    # via
    #   scipy
    #   torchvision
    #   transformers
packaging==23.1
    # via
    #   huggingface-hub
    #   transformers
pillow==9.5.0
    # via torchvision
protobuf==3.20.0
    # via -r ./solution/requirements/requirements.in
pydantic==1.10.7
    # via fastapi
python-dotenv==1.0.0
    # via -r ./solution/requirements/requirements.in
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2023.5.5
    # via transformers
requests==2.30.0
    # via
    #   huggingface-hub
    #   torchvision
    #   transformers
scipy==1.10.1
    # via -r ./solution/requirements/requirements.in
sentencepiece==0.1.99
    # via -r ./solution/requirements/requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
sympy==1.12
    # via torch
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   -r ./solution/requirements/requirements.in
    #   torchaudio
    #   torchvision
torchaudio==2.0.2
    # via -r ./solution/requirements/requirements.in
torchvision==0.15.2
    # via -r ./solution/requirements/requirements.in
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   transformers
transformers==4.29.0
    # via -r ./solution/requirements/requirements.in
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pydantic
    #   starlette
    #   torch
urllib3==2.0.2
    # via requests
uvicorn==0.22.0
    # via -r ./solution/requirements/requirements.in
